Moved `Val` from `ui_node.rs` into `geometry.rs`
